+++ packages/Webkul/Admin/src/Routes/rest-routes.php
@@ -24,6 +24,8 @@ Route::group(['middleware' => ['admin'], 'prefix' => config('app.admin_url')], f
 
     Route::get('media/download', [MediaController::class, 'download'])->name('admin.media.download');
 
+    Route::post('media/scan', [MediaController::class, 'scan'])->name('admin.media.scan');
+
     /**
      * Dashboard routes.
      */